+++ brathena/src/map/elemental.h
@@ -64,7 +64,7 @@ struct elemental_data {
 	int summon_timer;
 	int skill_timer;
 
-	unsigned last_thinktime, last_linktime, last_spdrain_time;
+	int64 last_thinktime, last_linktime, last_spdrain_time;
 	short min_chase;
 	int target_id, attacked_id;
 };
@@ -92,7 +92,7 @@ int elemental_skillnotok(uint16 skill_id, struct elemental_data *ed);
 int elemental_set_target(struct map_session_data *sd, struct block_list *bl);
 int elemental_clean_single_effect(struct elemental_data *ed, uint16 skill_id);
 int elemental_clean_effect(struct elemental_data *ed);
-int elemental_action(struct elemental_data *ed, struct block_list *bl, unsigned int tick);
+int elemental_action(struct elemental_data *ed, struct block_list *bl, int64 tick);
 struct skill_condition elemental_skill_get_requirements(uint16 skill_id, uint16 skill_lv);
 
 #define elemental_stop_walking(ed, type) unit_stop_walking(&(ed)->bl, type)
