+++ brathena/src/map/map.c
@@ -253,7 +253,7 @@ int map_freeblock_unlock(void)
 
 // Timer function to check if there some remaining lock and remove them if so.
 // Called each 1s
-int map_freeblock_timer(int tid, unsigned int tick, int id, intptr_t data)
+int map_freeblock_timer(int tid, int64 tick, int id, intptr_t data)
 {
 	if(block_free_lock > 0) {
 		ShowError("map_freeblock_timer: block_free_lock(%d) is invalid.\n", block_free_lock);
@@ -392,7 +392,7 @@ int map_delblock(struct block_list *bl)
  * Pass flag as 1 to prevent doing skill_unit_move checks
  * (which are executed by default on BL_CHAR types)
  *------------------------------------------*/
-int map_moveblock(struct block_list *bl, int x1, int y1, unsigned int tick)
+int map_moveblock(struct block_list *bl, int x1, int y1, int64 tick)
 {
 	int x0 = bl->x, y0 = bl->y;
 	struct status_change *sc = NULL;
@@ -1225,7 +1225,7 @@ int map_get_new_object_id(void)
  * Timered function to clear the floor (remove remaining item)
  * Called each flooritem_lifetime ms
  *------------------------------------------*/
-int map_clearflooritem_timer(int tid, unsigned int tick, int id, intptr_t data)
+int map_clearflooritem_timer(int tid, int64 tick, int id, intptr_t data)
 {
 	struct flooritem_data *fitem = (struct flooritem_data *)idb_get(id_db, id);
 
@@ -2161,7 +2161,7 @@ int map_removemobs_sub(struct block_list *bl, va_list ap)
 	return 1;
 }
 
-int map_removemobs_timer(int tid, unsigned int tick, int id, intptr_t data)
+int map_removemobs_timer(int tid, int64 tick, int id, intptr_t data)
 {
 	int count;
 	const int16 m = id;
