+++ brathena/src/map/pet.c
@@ -203,7 +203,7 @@ int pet_sc_check(struct map_session_data *sd, int type)
 	return 0;
 }
 
-static int pet_hungry(int tid, unsigned int tick, int id, intptr_t data)
+static int pet_hungry(int tid, int64 tick, int id, intptr_t data)
 {
 	struct map_session_data *sd;
 	struct pet_data *pd;
@@ -686,7 +686,7 @@ int pet_equipitem(struct map_session_data *sd,int index)
 	clif_send_petdata(NULL, sd->pd, 3, sd->pd->vd.head_bottom);
 	if(battle_config.pet_equip_required) {
 		//Skotlex: start support timers if need
-		unsigned int tick = gettick();
+		int64 tick = gettick();
 		if(pd->s_skill && pd->s_skill->timer == INVALID_TIMER) {
 			if(pd->s_skill->id)
 				pd->s_skill->timer=add_timer(tick+pd->s_skill->delay*1000, pet_skill_support_timer, sd->bl.id, 0);
@@ -785,7 +785,7 @@ static int pet_food(struct map_session_data *sd, struct pet_data *pd)
 	return 0;
 }
 
-static int pet_randomwalk(struct pet_data *pd,unsigned int tick)
+static int pet_randomwalk(struct pet_data *pd,int64 tick)
 {
 	nullpo_ret(pd);
 
@@ -826,7 +826,7 @@ static int pet_randomwalk(struct pet_data *pd,unsigned int tick)
 	return 0;
 }
 
-static int pet_ai_sub_hard(struct pet_data *pd, struct map_session_data *sd, unsigned int tick)
+static int pet_ai_sub_hard(struct pet_data *pd, struct map_session_data *sd, int64 tick)
 {
 	struct block_list *target = NULL;
 
@@ -939,14 +939,14 @@ static int pet_ai_sub_hard(struct pet_data *pd, struct map_session_data *sd, uns
 
 static int pet_ai_sub_foreachclient(struct map_session_data *sd,va_list ap)
 {
-	unsigned int tick = va_arg(ap,unsigned int);
+	int64 tick = va_arg(ap,int64);
 	if(sd->status.pet_id && sd->pd)
 		pet_ai_sub_hard(sd->pd,sd,tick);
 
 	return 0;
 }
 
-static int pet_ai_hard(int tid, unsigned int tick, int id, intptr_t data)
+static int pet_ai_hard(int tid, int64 tick, int id, intptr_t data)
 {
 	map_foreachpc(pet_ai_sub_foreachclient,tick);
 
@@ -979,7 +979,7 @@ static int pet_ai_sub_hard_lootsearch(struct block_list *bl,va_list ap)
 	return 0;
 }
 
-static int pet_delay_item_drop(int tid, unsigned int tick, int id, intptr_t data)
+static int pet_delay_item_drop(int tid, int64 tick, int id, intptr_t data)
 {
 	struct item_drop_list *list;
 	struct item_drop *ditem, *ditem_prev;
@@ -1047,7 +1047,7 @@ int pet_lootitem_drop(struct pet_data *pd,struct map_session_data *sd)
 /*==========================================
  * pet bonus giving skills [Valaris] / Rewritten by [Skotlex]
  *------------------------------------------*/
-int pet_skill_bonus_timer(int tid, unsigned int tick, int id, intptr_t data)
+int pet_skill_bonus_timer(int tid, int64 tick, int id, intptr_t data)
 {
 	struct map_session_data *sd=map_id2sd(id);
 	struct pet_data *pd;
@@ -1089,7 +1089,7 @@ int pet_skill_bonus_timer(int tid, unsigned int tick, int id, intptr_t data)
 /*==========================================
  * pet recovery skills [Valaris] / Rewritten by [Skotlex]
  *------------------------------------------*/
-int pet_recovery_timer(int tid, unsigned int tick, int id, intptr_t data)
+int pet_recovery_timer(int tid, int64 tick, int id, intptr_t data)
 {
 	struct map_session_data *sd=map_id2sd(id);
 	struct pet_data *pd;
@@ -1117,7 +1117,7 @@ int pet_recovery_timer(int tid, unsigned int tick, int id, intptr_t data)
 	return 0;
 }
 
-int pet_heal_timer(int tid, unsigned int tick, int id, intptr_t data)
+int pet_heal_timer(int tid, int64 tick, int id, intptr_t data)
 {
 	struct map_session_data *sd=map_id2sd(id);
 	struct status_data *status;
@@ -1155,7 +1155,7 @@ int pet_heal_timer(int tid, unsigned int tick, int id, intptr_t data)
 /*==========================================
  * pet support skills [Skotlex]
  *------------------------------------------*/
-int pet_skill_support_timer(int tid, unsigned int tick, int id, intptr_t data)
+int pet_skill_support_timer(int tid, int64 tick, int id, intptr_t data)
 {
 	struct map_session_data *sd=map_id2sd(id);
 	struct pet_data *pd;
